+++ Tests/Unit/AutoRouteManagerTest.php
@@ -25,13 +25,10 @@ class AutoRouteManagerTest extends \PHPUnit_Framework_TestCase
         $this->driver = $this->getMock('Symfony\Cmf\Component\RoutingAuto\AdapterInterface');
         $this->uriGenerator = $this->getMock('Symfony\Cmf\Component\RoutingAuto\UriGeneratorInterface');
         $this->defunctRouteHandler = $this->getMock('Symfony\Cmf\Component\RoutingAuto\DefunctRouteHandlerInterface');
-        $this->eventDispatcher = new EventDispatcher();
-        $this->eventDispatcher->addListener(RoutingAutoEvents::URI_CONTEXT_BUILT, array($this, 'onUriContextBuilt'));
         $this->autoRouteManager = new AutoRouteManager(
             $this->driver,
             $this->uriGenerator,
-            $this->defunctRouteHandler,
-            $this->eventDispatcher
+            $this->defunctRouteHandler
         );
     }
 
@@ -96,14 +93,5 @@ class AutoRouteManagerTest extends \PHPUnit_Framework_TestCase
         foreach ($expectedRoutes as $expectedRoute) {
             $this->assertTrue($uriContextCollection->containsAutoRoute($expectedRoute), 'URL context collection contains route: ' . spl_object_hash($expectedRoute));
         }
-    }
-
-    /**
-     * For testing the event listener
-     */
-    public function onUriContextBuilt(UriContextEvent $event)
-    {
-        $uriContext = $event->getUriContext();
-        $this->assertInstanceOf('Symfony\Cmf\Component\RoutingAuto\UriContext', $uriContext);
     }
 }